requirements.txt
pyopencl
numpy
scipy
pillow
matplotlib
opencv-python
pylabs
snakeviz
tqdm